Added log out functionality
