+++ blog_app_sqa/src/app/components/CreatePost.js
@@ -22,7 +22,7 @@ function CreatePost() {
     });
 
     if (response.ok) {
-      router.push("/");
+      router.push("/dashboard");
       router.refresh();
     }
   };
